Remove lmdb

+++ packages/db/clibs/include/selva/sort.h
@@ -74,7 +74,7 @@ SELVA_EXPORT
 void selva_sort_remove_i64(struct SelvaSortCtx *ctx, int64_t v, const void *p);
 
 SELVA_EXPORT
-void selva_sort_remove_double(struct SelvaSortCtx *ctx, int64_t d, const void *p);
+void selva_sort_remove_double(struct SelvaSortCtx *ctx, double d, const void *p);
 
 SELVA_EXPORT
 void selva_sort_remove_buf(struct SelvaSortCtx *ctx, const void *buf, size_t len, const void *p);
